pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pci-auditor"
version = "1.0.0"
description = "PCI DSS 4.0 compliance auditor for codebases and pull requests"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "openai>=1.30",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.7",
    "gitpython>=3.1",
    "sarif-om>=1.0.4",
    "jschema-to-python>=1.2.3",
    "colorama>=0.4.6",
    "python-dotenv>=1.0",
]

[project.scripts]
pci-auditor = "pci_auditor.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "responses>=0.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pci_auditor*"]

[tool.setuptools.package-data]
pci_auditor = ["rules/pci_rules.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
